+++ Sources/Screens/Loading/LoginLoadingViewModel.swift
@@ -14,9 +14,9 @@ struct LoadingViewModel: GDSScreenViewModel, BaseViewModel {
     let backButtonTitle: GDSLocalisedString? = nil
     let backButtonIsHidden: Bool = true
     
+    let loadingLabelKey: GDSLocalisedString = "app_loadingBody"
     var didAppear: DesignSystem.Action? {
         .action {
-            let loadingLabelKey: GDSLocalisedString = "app_loadingBody"
             let screen = ScreenView(id: IntroAnalyticsScreenID.loginLoading.rawValue,
                                     screen: IntroAnalyticsScreen.loginLoading,
                                     titleKey: loadingLabelKey.stringKey)
